requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib
python-jose[cryptography]
python-multipart
pydantic>=2.0.0